switch to two types of slack tasks
