+++ services/QuillLMS/Rakefile
@@ -5,6 +5,7 @@
 
 require File.expand_path('../config/application', __FILE__)
 require 'dotenv/tasks'
+require 'tasks/modules/slack_tasks'
 
 EmpiricalGrammar::Application.load_tasks
